Refactor names and locations of files. Driver: Carlos, Navigator: Jax
